Match mutatedEbaUpdateWhileFrozen via (int) compare casts
Cast each comparand to (int) in the move==0||==1||==3||==4 chain to
force signed cmpwi equality tests and defeat MWCC's unsigned
adjacent-value range-fold (==0||==1 -> clrlwi;cmplwi;ble). 95.20 -> 100%.

+++ src/main/dll/duster.c
@@ -903,7 +903,7 @@ void mutatedEbaUpdateWhileFrozen(uint obj, int state, undefined4 unused, int eve
         else
         {
             move = ((GameObject*)obj)->anim.currentMove;
-            if ((((move == 0) || (move == 1)) || (move == 3)) || (move == 4))
+            if ((int)move == 0 || (int)move == 1 || (int)move == 3 || (int)move == 4)
             {
                 Sfx_PlayFromObject(obj, SFXfox_roll2);
                 ((BaddieState*)state)->reactionFlags = ((BaddieState*)state)->reactionFlags | 0x10;
